Move spreadsheet health test.
